Seed instances from bundled gitignored file; seed previews
Now that the project uses synchronized folders, replace the runtime
Documents-directory read with a bundled `seed-instances.json` living in
`Ruddarr/Preview Content/` (a development asset, so it's debug-only and
stripped from Release). The real file is gitignored — absent on a fresh
clone the build simply skips it — so credentials never enter the repo.

- Move the seed DTO + upsert logic into its own InstanceSeeder.swift
  (auto-included via the synchronized folder, no project edits).
- Read the seed file from the app bundle instead of Documents; works in
  both the running app and Xcode Previews.
- Seed the Settings preview with the existing dummy fixtures.
- Relocate seed-instances.example.json next to the real file.
- Delete scripts/seed-instances.sh — no longer needed.
